+++ src/java/controller/RemoveItemCart.java
@@ -33,7 +33,7 @@ public class RemoveItemCart extends HttpServlet {
 		String customerId = user.getId();
 		String productId = request.getParameter("product_id");
 		String url = request.getParameter("goto");
-		int size = Integer.parseInt(request.getParameter("size"));
+		float size = Float.parseFloat(request.getParameter("size"));
 		CartDAO cartAccess = new CartDAO();
 		boolean r = cartAccess.removeItem(customerId, productId, size);
 		if(!r){
