+++ src/java/entity/ProductInCart.java
@@ -6,18 +6,18 @@ package entity;
  */
 public class ProductInCart extends Product {
 	private int quantity;
-	private int size;
+	private float size;
 
 	public ProductInCart() {
 	}
 
-	public ProductInCart(int quantity, int size, String id, String name, String brand, double price, int unitsInStock, int orderLevel, int view, double discount, String description, String picture) {
+	public ProductInCart(int quantity, float size, String id, String name, String brand, double price, int unitsInStock, int orderLevel, int view, double discount, String description, String picture) {
 		super(id, name, brand, price, unitsInStock, orderLevel, view, discount, description, picture);
 		this.quantity = quantity;
 		this.size = size;
 	}
 
-	public int getSize() {
+	public float getSize() {
 		return size;
 	}
